requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.12.5=ha878542_0
cairo=1.14.12=he6fea26_5
certifi=2020.12.5=py37h89c1867_1
cudatoolkit=10.2.89=hfd86e86_1
faiss=1.7.0=py37cuda102h6f2cbd8_8_cuda
faiss-gpu=1.6.3=py37h1a5d453_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=he4413a7_1000
freeglut=3.2.1=h58526e2_0
freetype=2.10.4=h5ab3b9f_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.55.0=0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
graphite2=1.3.13=h58526e2_1001
harfbuzz=1.9.0=h08d66d9_0
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
intel-openmp=2021.2.0=h06a4308_610
jasper=2.0.14=h07fcdf6_1
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.9.0=9_mkl
libfaiss=1.7.0=cuda102hd37495c_8_cuda
libfaiss-avx2=1.7.0=cuda102h1234567_8_cuda
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=he1b5a44_1001
libgomp=9.3.0=h2828fa1_19
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
liblapack=3.9.0=9_mkl
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.9=h13577e0_2
lz4-c=1.9.3=h2531618_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py37h27cfd23_1
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.1=py37ha9443f7_2
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.2=py37h2d18471_0
numpy-base=1.20.2=py37hfae3a4d_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h7f98852_0
pcre=8.44=he1b5a44_0
pillow=8.2.0=py37he98fc37_0
pip=21.1.1=py37h06a4308_0
pixman=0.34.0=h14c3975_1003
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=3.4.2=py37hb342d67_1
python=3.7.10=hdb3f193_0
python_abi=3.7=1_cp37m
pytorch=1.8.1=py3.7_cuda10.2_cudnn7.6.5_0
readline=8.1=h27cfd23_0
scikit-learn=0.24.2=py37ha9443f7_0
scipy=1.6.2=py37had2a1c9_1
setuptools=52.0.0=py37h06a4308_0
six=1.15.0=py37h06a4308_0
sqlite=3.35.4=hdfb4753_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.2.1=py37_0
tqdm=4.59.0=pyhd3eb1b0_1
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.1=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
